Add EntranceTank class; refactor floc and sed generally
