+++ aguaclara/design/ent_tank.py
@@ -1,13 +1,14 @@
 from aguaclara.core.units import unit_registry as u
-from core.units import unit_registry as u
+
+from design.floc import L_SED, Flocculator
 
 L_MAX = 2.2 * u.m
 
 # Angle of the sloped walls of the entrance tank hoppers
 ENT_TANK_SLOPE_ANGLE = 45 * u.deg
 
-# Extra space around the float (increase in effective diameter) to ensure
-# that float has free travel
+# Extra space around the float (increase in effective diameter) to
+# ensure that float has free travel
 FLOAT_S = 5 * u.cm
 HOPPER_PEAK_W = 3 * u.cm
 PLATE_S = 2.5 * u.cm
@@ -39,16 +40,17 @@ RAPID_MIX_EDR = 3 * u.W / u.kg
 
 RAPID_MIX_PLATE_RESTRAINER_ND = 0.5 * u.inch
 
-FLOAT_ND =  8*u.inch
+FLOAT_ND = 8*u.inch
 
-#Minimum pipe size to handle grit and to ensure that the pipe can be easily unclogged
-DRAIN_MIN_ND =  3*u.inch
+# Minimum pipe size to handle grit and to ensure that the pipe can be
+# easily unclogged
+DRAIN_MIN_ND = 3*u.inch
 
-DRAIN_ND =  3*u.inch #This is constant for now
+DRAIN_ND = 3*u.inch  # This is constant for now
 
 REMOVABLE_WALL_THICKNESS = 5*u.cm
 
-#Parameters are arbitrary - need to be calculated
+# Parameters are arbitrary - need to be calculated
 REMOVABLE_WALL_SUPPORT_H = 4 * u.cm
 
 REMOVABLE_WALL_SUPPORT_THICKNESS = 5*u.cm
@@ -57,3 +59,39 @@ HOPPER_LEDGE_THICKNESS = 15*u.cm
 WALKWAY_W = 1 * u.m
 RAPID_MIX_ORIFICE_PLATE_THICKNESS = 2*u.cm
 RAPID_MIX_AIR_RELEASE_ND = 1*u.inch
+
+
+class EntranceTank:
+
+    def __init__(self, floc=Flocculator()):
+        self.floc = floc
+        self.q = floc.q
+        self.temp = floc.temp
+
+    def a_ent_tank(self):
+        """Return the planview area of the entrance tank given plant flow rate,
+        headloss, target collision potential, design temperature, and depth of
+        water at the end of the flocculator.
+
+        Examples
+        --------
+        >>> from aguaclara.play import *
+        >>> area_ent_tank(20*u.L/u.s, 40*u.cm, 37000, 25*u.degC, 2*u.m)
+        1 meter ** 2
+        """
+        # first guess planview area
+        a_new = 1 * u.m**2
+        a_ratio = 0
+        tolerance = 0.01
+        while a_ratio > (1 + tolerance):
+            a_et_pv = a_new
+            a_floc_pv = self.floc.vol / (self.floc.END_WATER_HEIGHT + (
+                self.floc.HL / 2))
+            a_etf_pv = a_et_pv + a_floc_pv
+
+            w_tot = a_etf_pv / L_SED
+            w_chan = w_tot / self.floc.num_channel
+
+            a_new = L_MAX * w_chan
+            a_ratio = a_new / a_et_pv
+        return a_new
